Attendance Module Authorized !!

+++ app/Http/Controllers/AttendanceController.php
@@ -19,43 +19,51 @@ class AttendanceController extends Controller
      */
     public function index()
     {
-        return view('hrm.attendance.index');
+        if (auth()->user()->can('attendance_view_all') || auth()->user()->can('attendance_view_own')) {
+            return view('hrm.attendance.index');
+        }
     }
 
     public function create()
     {
-        $company = DB::table('company')->get()->all();
-        $employees = Employee::get()->all();
-        $offices = Office::get()->all();
-        return view('hrm.attendance.create', compact('company', 'employees', 'offices'));
+        if (auth()->user()->can('attendance_create')) {
+            $company = DB::table('company')->get()->all();
+            $employees = Employee::get()->all();
+            $offices = Office::get()->all();
+            return view('hrm.attendance.create', compact('company', 'employees', 'offices'));
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'company' => 'required|exists:company,id',
-            'employee' => 'required|exists:employee_shift,id',
-            'shift_name' => 'required|exists:office_shift,id',
-            'date' => 'required|date',
-            'clock_in' => 'required|date_format:H:i',
-            // 'clock_out' => 'required|date_format:H:i|after:clock_in',
-            'clock_out' => 'required|date_format:H:i|after:clock_in',
-        ]);
+        if (auth()->user()->can('attendance_create')) {
+            $request->validate([
+                'company' => 'required|exists:company,id',
+                'employee' => 'required|exists:employee_shift,id',
+                'shift_name' => 'required|exists:office_shift,id',
+                'date' => 'required|date',
+                'clock_in' => 'required|date_format:H:i',
+                // 'clock_out' => 'required|date_format:H:i|after:clock_in',
+                'clock_out' => 'required|date_format:H:i|after:clock_in',
+            ]);
 
-        // Create a new Attendance model and fill it with the form data
-        $attendance = new Attendance();
-        $attendance->company_id = $request->input('company');
-        $attendance->emp_id = $request->input('employee');
-        $attendance->office_id = $request->input('shift_name');
-        $attendance->date = $request->input('date');
-        $attendance->clock_in = $request->input('clock_in');
-        $attendance->clock_out = $request->input('clock_out');
+            // Create a new Attendance model and fill it with the form data
+            $attendance = new Attendance();
+            $attendance->company_id = $request->input('company');
+            $attendance->emp_id = $request->input('employee');
+            $attendance->office_id = $request->input('shift_name');
+            $attendance->date = $request->input('date');
+            $attendance->clock_in = $request->input('clock_in');
+            $attendance->clock_out = $request->input('clock_out');
 
-        // Save the attendance record
-        $attendance->save();
+            // Save the attendance record
+            $attendance->save();
 
-        // Redirect to the index page or do any other necessary actions
-        return redirect()->route('attendance.index')->with('success', 'Attendance record created successfully');
+            // Redirect to the index page or do any other necessary actions
+            return redirect()->route('attendance.index')->with('success', 'Attendance record created successfully');
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     public function getData()
@@ -95,55 +103,75 @@ class AttendanceController extends Controller
                 // Add your condition on the employee table here
                 $query->where('user_id', auth()->user()->id);
             })
-            ->with(['company', 'employee', 'office'])->get();
+                ->with(['company', 'employee', 'office'])->get();
             return response()->json(['data' => $attendance]);
         }
+        return abort('403', __('You are not authorized'));
     }
 
 
     public function edit($id)
     {
-        $company = Company::all();
-        $employees = Employee::all();
-        $offices = Office::all();
+        if (auth()->user()->can('attendance_edit') && auth()->user()->can('attendance_view_all')) {
+            $company = Company::all();
+            $employees = Employee::all();
+            $offices = Office::all();
+            $attendance = Attendance::with(['employee', 'company', 'office'])->findOrFail($id);
+            return view('hrm.attendance.edit', compact('attendance', 'company', 'employees', 'offices'));
+        }
+        if (auth()->user()->can('attendance_edit') && auth()->user()->can('attendance_view_own')) {
+            $company = Company::all();
+            $employees = Employee::all();
+            $offices = Office::all();
+            $attendance = Attendance::with(['employee', 'company', 'office'])->findOrFail($id);
+            if($attendance->employee->user_id != auth()->user()->id){
+                return abort('403', __('You are not authorized'));
+            }
+            return view('hrm.attendance.edit', compact('attendance', 'company', 'employees', 'offices'));
+        }
 
-        $attendance = Attendance::with(['employee', 'company', 'office'])->findOrFail($id);
-
-        return view('hrm.attendance.edit', compact('attendance', 'company', 'employees', 'offices'));
+        return abort('403', __('You are not authorized'));
     }
 
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'company' => 'required|exists:company,id',
-            'employee' => 'required|exists:employee_shift,id',
-            'shift_name' => 'required|exists:office_shift,id',
-            'date' => 'required|date',
-            'clock_in' => 'required|date_format:H:i',
-            'clock_out' => 'required|date_format:H:i|after:clock_in',
-        ]);
+        if (auth()->user()->can('attendance_edit')) {
+            $request->validate([
+                'company' => 'required|exists:company,id',
+                'employee' => 'required|exists:employee_shift,id',
+                'shift_name' => 'required|exists:office_shift,id',
+                'date' => 'required|date',
+                'clock_in' => 'required|date_format:H:i',
+                'clock_out' => 'required|date_format:H:i|after:clock_in',
+            ]);
 
-        $attendance = Attendance::findOrFail($id);
-        $attendance->company_id = $request->input('company');
-        $attendance->emp_id = $request->input('employee');
-        $attendance->office_id = $request->input('shift_name');
-        $attendance->date = $request->input('date');
-        $attendance->clock_in = $request->input('clock_in');
-        $attendance->clock_out = $request->input('clock_out');
+            $attendance = Attendance::findOrFail($id);
+            $attendance->company_id = $request->input('company');
+            $attendance->emp_id = $request->input('employee');
+            $attendance->office_id = $request->input('shift_name');
+            $attendance->date = $request->input('date');
+            $attendance->clock_in = $request->input('clock_in');
+            $attendance->clock_out = $request->input('clock_out');
 
-        $attendance->save();
+            $attendance->save();
 
-        return redirect()->route('attendance.index')->with('success', 'Attendance record updated successfully');
+            return redirect()->route('attendance.index')->with('success', 'Attendance record updated successfully');
+        }
+        return abort('403', __('You are not authorized'));
     }
     public function delete(Request $request)
     {
-        $attendance = Attendance::findOrFail($request->id);
+        if (auth()->user()->can('attendance_delete')) {
 
-        if ($attendance) {
-            $attendance->delete();
-            return response()->json(['message' => 'Attendance deleted successfully'], 200);
-        } else {
-            return response()->json(['message' => 'Attendance not found'], 404);
+            $attendance = Attendance::findOrFail($request->id);
+
+            if ($attendance) {
+                $attendance->delete();
+                return response()->json(['message' => 'Attendance deleted successfully'], 200);
+            } else {
+                return response()->json(['message' => 'Attendance not found'], 404);
+            }
         }
+        return abort('403', __('You are not authorized'));
     }
 }
